Refactor shared *Manager interfaces into provider package
SecretsManager, ConfigMapsManager, DataVolumesManager, and
VirtualMachineManager can all be used by the Vmware provider.
This pulls those interfaces out of the oVirt package
and into the root provider package.
