+++ pkg/providers/provider.go
@@ -7,6 +7,7 @@ import (
 	kubevirtv1 "kubevirt.io/client-go/api/v1"
 	cdiv1 "kubevirt.io/containerized-data-importer/pkg/apis/core/v1alpha1"
 	rclient "sigs.k8s.io/controller-runtime/pkg/client"
+	"k8s.io/apimachinery/pkg/types"
 )
 
 const (
@@ -46,3 +47,29 @@ type Mapper interface {
 
 // VMStatus represents VM status
 type VMStatus string
+
+// SecretsManager defines operations on secrets
+type SecretsManager interface {
+	FindFor(types.NamespacedName) (*corev1.Secret, error)
+	CreateFor(*corev1.Secret, types.NamespacedName) error
+	DeleteFor(types.NamespacedName) error
+}
+
+// ConfigMapsManager defines operations on config maps
+type ConfigMapsManager interface {
+	FindFor(types.NamespacedName) (*corev1.ConfigMap, error)
+	CreateFor(*corev1.ConfigMap, types.NamespacedName) error
+	DeleteFor(types.NamespacedName) error
+}
+
+// DataVolumesManager defines operations on datavolumes
+type DataVolumesManager interface {
+	FindFor(types.NamespacedName) ([]*cdiv1.DataVolume, error)
+	DeleteFor(types.NamespacedName) error
+}
+
+// VirtualMachineManager defines operations on datavolumes
+type VirtualMachineManager interface {
+	FindFor(types.NamespacedName) (*kubevirtv1.VirtualMachine, error)
+	DeleteFor(types.NamespacedName) error
+}
